Refactor Civic Relevance in new file
