+++ src/civic/index.js
@@ -19,6 +19,7 @@ const _pubmed = require('../entrez/pubmed');
 const _entrezGene = require('../entrez/gene');
 const { civic: SOURCE_DEFN, ncit: NCIT_SOURCE_DEFN } = require('../sources');
 const { processVariantRecord } = require('./variant');
+const { getRelevance } = require('./relevance');
 const { getPublication } = require('./publication');
 const { EvidenceItem: evidenceSpec } = require('./specs.json');
 
@@ -78,351 +79,6 @@ const requestEvidenceItems = async (url, opt) => {
     return allRecords;
 };
 
-
-/**
- * Extract the appropriate GraphKB relevance term from a CIViC evidence record
- */
-const translateRelevance = (evidenceType, evidenceDirection, significance) => {
-    if (evidenceDirection === 'DOES_NOT_SUPPORT') {
-        switch (evidenceType) { // eslint-disable-line default-case
-            case 'DIAGNOSTIC': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'NEGATIVE': {
-                        // 2 cases
-                        break;
-                    }
-
-                    case 'POSITIVE': {
-                        // 11 cases
-                        break;
-                    }
-                }
-            }
-
-            case 'FUNCTIONAL': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'DOMINANT_NEGATIVE': {
-                        // 83 cases to address
-                        break;
-                    }
-
-                    case 'GAIN_OF_FUNCTION': {
-                        // 11 cases to address
-                        break;
-                    }
-
-                    case 'LOSS_OF_FUNCTION': {
-                        // 2 cases to address
-                        break;
-                    }
-
-                    case 'NEOMORPHIC': {
-                        // 6 cases
-                        break;
-                    }
-
-                    case 'UNALTERED_FUNCTION': {
-                        // No case so far
-                        break;
-                    }
-
-                    case 'UNKNOWNED': {
-                        // 1 case (to address?)
-                        break;
-                    }
-                }
-            }
-
-            case 'ONCOGENIC': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'ONCOGENICITY': {
-                        // 3 cases to address
-                        break;
-                    }
-
-                    case 'PROTECTIVENESS': {
-                        // No cases so far
-                        break;
-                    }
-                }
-            }
-
-            case 'PREDICTIVE': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'ADVERSE_RESPONSE': {
-                        // 1 case to address
-                        break;
-                    }
-
-                    case 'NA': {
-                        // 2 cases
-                        // According to https://civic.readthedocs.io/en/latest/model/evidence/significance.html
-                        // "Variant does not inform clinical interepretation"
-                        break;
-                    }
-
-                    case 'REDUCED_SENSITIVITY': {
-                        // 2 cases to address
-                        break;
-                    }
-
-                    case 'RESISTANCE': {
-                        return 'no resistance';
-                    }
-
-                    case 'SENSITIVITYRESPONSE': {
-                        return 'no response';
-                    }
-                }
-            }
-
-            case 'PREDISPOSING': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'LIKELY_PATHOGENIC': {
-                        // 1 case to address. Should not append according to CIViC docs
-                        break;
-                    }
-
-                    case 'POSITIVE': {
-                        // 1 case. Should not append according to CIViC docs
-                        break;
-                    }
-
-                    case 'PREDISPOSITION': {
-                        // 2 cases to address
-                        break;
-                    }
-
-                    case 'PROTECTIVENESS': {
-                        // No case so far
-                        break;
-                    }
-                }
-            }
-
-            case 'PROGNOSTIC': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'BETTER_OUTCOME': {
-                        // 4 case to address.
-                        break;
-                    }
-
-                    case 'NA': {
-                        // 38 cases (to address?)
-                        // According to https://civic.readthedocs.io/en/latest/model/evidence/significance.html
-                        // "The N/A option can be used to imply that the variant does not have an impact on patient prognosis."
-                        break;
-                    }
-
-                    case 'POOR_OUTCOME': {
-                        // 34 cases
-                        break;
-                    }
-                }
-            }
-        }
-    } else if (evidenceDirection === 'SUPPORTS') {
-        switch (evidenceType) { // eslint-disable-line default-case
-            case 'DIAGNOSTIC': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'NA': {
-                        // 9 cases (to address? 'POSITIVE' implied?). Should not append according to CIViC docs
-                        break;
-                    }
-
-                    case 'NEGATIVE': {
-                        return 'opposes diagnosis';
-                    }
-
-                    case 'POSITIVE': {
-                        return 'favours diagnosis';
-                    }
-                }
-            }
-
-            case 'FUNCTIONAL': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'DOMINANT_NEGATIVE': {
-                        return 'dominant negative';
-                    }
-
-                    case 'GAIN_OF_FUNCTION': {
-                        return 'gain of function';
-                    }
-
-                    case 'LOSS_OF_FUNCTION': {
-                        return 'loss of function';
-                    }
-
-                    case 'NEOMORPHIC': {
-                        return 'neomorphic';
-                    }
-
-                    case 'UNALTERED_FUNCTION': {
-                        return 'unaltered function';
-                    }
-
-                    case 'UNKNOWNED': {
-                        // No case so far
-                        break;
-                    }
-                }
-            }
-
-            case 'ONCOGENIC': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'ONCOGENICITY': {
-                        // 92 cases.
-                        // So far, only OncoKB and IPRKB had statement of relevance 'oncogenic'.
-                        // Needs confirmation !!
-                        return 'oncogenic';
-                    }
-
-                    case 'PROTECTIVENESS': {
-                        // No case so far
-                        break;
-                    }
-                }
-            }
-
-            case 'PREDICTIVE': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'ADVERSE_RESPONSE': {
-                        return 'adverse response';
-                    }
-
-                    case 'NA': {
-                        // According to https://civic.readthedocs.io/en/latest/model/evidence/significance.html
-                        // "Variant does not inform clinical interepretation"
-                        break;
-                    }
-
-                    case 'REDUCED_SENSITIVITY': {
-                        return 'reduced sensitivity';
-                    }
-
-                    case 'RESISTANCE': {
-                        return 'resistance';
-                    }
-
-                    case 'SENSITIVITYRESPONSE': {
-                        return 'sensitivity';
-                    }
-                }
-            }
-
-            case 'PREDISPOSING': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'LIKELY_PATHOGENIC': {
-                        // 23 cases. Should not append according to CIViC docs
-                        return 'likely pathogenic';
-                    }
-
-                    case 'NA': {
-                        // 3 cases. Should not append according to CIViC docs
-                        // Similar to 'UNCERTAIN_SIGNIFICANCE' ?
-                        // Needs confirmation !!
-                        return 'likely predisposing';
-                    }
-
-                    case 'PATHOGENIC': {
-                        // 13 cases. Should not append according to CIViC docs
-                        return 'pathogenic';
-                    }
-
-                    case 'POSITIVE': {
-                        // 3 cases. Should not append according to CIViC docs
-                        return 'predisposing';
-                    }
-
-                    case 'PREDISPOSITION': {
-                        return 'predisposing';
-                    }
-
-                    case 'PROTECTIVENESS': {
-                        // No case so far
-                        break;
-                    }
-
-                    case 'UNCERTAIN_SIGNIFICANCE': {
-                        // 1267 cases. Should not append according to CIViC docs
-                        return 'likely predisposing';
-                    }
-                }
-            }
-
-            case 'PROGNOSTIC': {
-                switch (significance) { // eslint-disable-line default-case
-                    case 'BETTER_OUTCOME': {
-                        return 'favourable prognosis';
-                    }
-
-                    case 'NA': {
-                        // 9 cases (to address?)
-                        // According to https://civic.readthedocs.io/en/latest/model/evidence/significance.html
-                        // "The N/A option can be used to imply that the variant does not have an impact on patient prognosis."
-                        break;
-                    }
-
-                    case 'POOR_OUTCOME': {
-                        return 'unfavourable prognosis';
-                    }
-
-                    case 'POSITIVE': {
-                        // 1 case. Should not append according to CIViC docs
-                        return 'favourable prognosis';
-                    }
-                }
-            }
-        }
-    }
-
-    // Addressing some NAs combinations
-    if (evidenceDirection === 'NA' && significance === 'NA') {
-        switch (evidenceType) { // eslint-disable-line default-case
-            case 'PREDISPOSING': {
-                // 1116 cases. Should not append according to CIViC docs
-                // Similar to 'SUPPORT'+'PREDISPOSING'+'UNCERTAIN_SIGNIFICANCE' ?
-                return 'likely predisposing'; // Needs confirmation !! Used to be tested as an error
-            }
-
-            case 'ONCOGENIC': {
-                // 166 cases (to address?)
-                // Similar to 'SUPPORT'+'ONCOGENIC'+'ONCOGENICITY' ?
-                // Needs confirmation !!
-                return 'oncogenic';
-            }
-        }
-    }
-
-    // If combination of evidenceDirection, evidenceType and significance not supported
-    throw new NotImplementedError(
-        `unable to process relevance (${JSON.stringify(
-            { evidenceDirection, evidenceType, significance },
-        )})`,
-    );
-};
-
-
-/**
- * Convert the CIViC relevance types to GraphKB terms
- */
-const getRelevance = async ({ rawRecord, conn }) => {
-    // translate the type to a GraphKB vocabulary term
-    let relevance = translateRelevance(
-        rawRecord.evidenceType,
-        rawRecord.evidenceDirection,
-        rawRecord.significance,
-    ).toLowerCase();
-
-    if (RELEVANCE_CACHE[relevance] === undefined) {
-        relevance = await conn.getVocabularyTerm(relevance);
-        RELEVANCE_CACHE[relevance.name] = relevance;
-    } else {
-        relevance = RELEVANCE_CACHE[relevance];
-    }
-    return relevance;
-};
 
 /**
  * Given some therapy name, find the therapy that is equivalent by name in GraphKB
@@ -1080,9 +736,8 @@ const upload = async ({
 
 
 module.exports = {
-    disambiguateVariant,
     SOURCE_DEFN,
+    disambiguateVariant,
     specs: { validateEvidenceSpec },
-    translateRelevance,
     upload,
 };
